requirements.txt
lyricsgenius
lxml
openai
langchain
pinecone-client
streamlit
tiktoken